Cleanup and pass test
